ai-model/requirements.txt
# ai-model/requirements.txt (업데이트)
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
pillow